+++ tests/test_utils/test_dicom.py
@@ -1,36 +1,13 @@
 import os
-import tempfile
 from unittest.mock import patch
 
 import numpy as np
 import nibabel as nib
 import pytest
 from nibabel.filebasedimages import ImageFileError
-from nibabel.spatialimages import HeaderDataError
 
 from redbrick.utils import dicom
 
-
-@pytest.fixture
-def input_nifti_file():
-    # Create a temporary NIfTI file for testing
-    data = np.array([[1, 1], [0, 0]])
-    img = nib.Nifti1Image(data, np.eye(4), dtype='compat')
-    with tempfile.NamedTemporaryFile(suffix=".nii.gz", delete=False) as f:
-        nib.save(img, f.name)
-        yield f.name
-    os.remove(f.name)
-
-
-@pytest.fixture
-def output_nifti_file():
-    # Create a temporary NIfTI file for testing
-    data = np.array([[2, 2], [0, 0]])
-    img = nib.Nifti1Image(data, np.eye(4), dtype='compat')
-    with tempfile.NamedTemporaryFile(suffix=".nii.gz", delete=False) as f:
-        nib.save(img, f.name)
-        yield f.name
-    os.remove(f.name)
 
 @pytest.mark.parametrize(
     ("equals", "pass_output", "expected"),
@@ -78,4 +55,3 @@ def test_merge_segmentations_invalid_nifti_file(input_nifti_file, output_nifti_f
         dicom.merge_segmentations(input_nifti_file, input_instance, equals, output_nifti_file, output_instance)
         exception = mock_logger.call_args[0][0]
         raise exception
-
